refactor: initCarousel 함수를 별도 파일로 분리
